Add raftgendis script to ctests
